backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
anthropic
composio-core
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
pydantic-settings
httpx
